requirements.txt
cffi==2.0.0
cryptography==46.0.3
customtkinter==5.2.2
darkdetect==0.8.0
packaging==25.0
pillow==11.3.0
pycparser==2.23
ttkbootstrap==1.18.1
